+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.6.6](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.5...@rudderstack/analytics-js-plugins@3.6.6) (2024-11-21)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.2`
 ## [3.6.5](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.4...@rudderstack/analytics-js-plugins@3.6.5) (2024-11-21)
 
 ### Dependency Updates