+++ packages/sanity-suite/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.1.37](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-sanity-suite@3.1.36...@rudderstack/analytics-js-sanity-suite@3.1.37) (2024-11-21)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.11.3`
 ## [3.1.36](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-sanity-suite@3.1.35...@rudderstack/analytics-js-sanity-suite@3.1.36) (2024-11-21)
 
 ### Dependency Updates